# frontend/src/types/openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "人生草稿本 API",
    "description": "Life Draft Book — Interactive narrative game API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/games": {
      "post": {
        "tags": [
          "Games"
        ],
        "summary": "Create Game",
        "description": "Create a new game from character settings.",
        "operationId": "create_game_api_games_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Games"
        ],
        "summary": "List Games",
        "description": "List saved games for the current user.",
        "operationId": "list_games_api_games_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameListItem"
                  },
                  "title": "Response List Games Api Games Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/presets": {
      "post": {
        "tags": [
          "Presets"
        ],
        "summary": "Create Preset",
        "description": "Save a character preset.",
        "operationId": "create_preset_api_presets_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePresetRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Presets"
        ],
        "summary": "List Presets",
        "description": "List character presets for the current user.",
        "operationId": "list_presets_api_presets_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PresetInfo"
                  },
                  "title": "Response List Presets Api Presets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/client-log": {
      "post": {
        "summary": "Client Log",
        "description": "Receive and log client-side errors — useful for debugging mobile issues.",
        "operationId": "client_log_api_client_log_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientLogEntry"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ClientLogEntry": {
        "properties": {
          "level": {
            "type": "string",
            "title": "Level",
            "default": "error"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "ua": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ua"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ClientLogEntry"
      },
...
      "CreateGameRequest": {
        "properties": {
          "character_settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Character Settings",
            "x-maxBytes": 262144
          },
          "player_name": {
            "type": "string",
            "maxLength": 50,
            "title": "Player Name"
          },
          "life_vision": {
            "type": "string",
            "maxLength": 500,
            "title": "Life Vision"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "zh"
          },
          "constraint_level": {
            "type": "string",
            "title": "Constraint Level",
            "default": "expert"
          }
        },
        "type": "object",
        "required": [
          "character_settings",
          "player_name",
          "life_vision"
        ],
        "title": "CreateGameRequest"
      },
      "CreatePresetRequest": {
        "properties": {
          "preset_name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Preset Name"
          },
          "player_name": {
            "type": "string",
            "maxLength": 50,
            "title": "Player Name"
          },
          "life_vision": {
            "type": "string",
            "maxLength": 500,
            "title": "Life Vision",
            "default": ""
          },
          "character_settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Character Settings",
            "x-maxBytes": 262144
          }
        },
        "type": "object",
        "required": [
          "preset_name",
          "player_name"
        ],
        "title": "CreatePresetRequest"
      },
...
      "GameListItem": {
        "properties": {
          "game_id": {
            "type": "integer",
            "title": "Game Id"
          },
          "player_name": {
            "type": "string",
            "title": "Player Name"
          },
          "week": {
            "type": "integer",
            "title": "Week"
          },
          "age": {
            "type": "integer",
            "title": "Age"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "has_progress": {
            "type": "boolean",
            "title": "Has Progress",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "player_name",
          "week",
          "age"
        ],
        "title": "GameListItem"
      },
      "GameStateResponse": {
        "properties": {
          "game_id": {
            "type": "integer",
            "title": "Game Id"
          },
          "player_state": {
            "additionalProperties": true,
            "type": "object",
            "title": "Player State"
          },
          "progress": {
            "additionalProperties": true,
            "type": "object",
            "title": "Progress"
          },
          "round_info": {
            "additionalProperties": true,
            "type": "object",
            "title": "Round Info"
          },
          "current_event": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Event"
          },
          "timeline": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeline"
          },
          "constraint_level": {
            "type": "string",
            "title": "Constraint Level",
            "default": "expert"
          },
          "narrative_style_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Narrative Style Id"
          },
          "narrative_style_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Narrative Style Name"
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "player_state",
          "progress",
          "round_info"
        ],
        "title": "GameStateResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PresetInfo": {
        "properties": {
          "preset_id": {
            "type": "integer",
            "title": "Preset Id"
          },
          "preset_name": {
            "type": "string",
            "title": "Preset Name"
          },
          "player_name": {
            "type": "string",
            "title": "Player Name"
          },
          "life_vision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Life Vision"
          },
          "character_settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Character Settings"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "preset_id",
          "preset_name",
          "player_name"
        ],
        "title": "PresetInfo"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}